Ignore unsupported signals.
SIGINFO does not exist on Linux (except on Alpha apparently). Cli-kit tries to
install a handler for this signal, which is the correct thing to do on BSD or
Darwin, but just raises on Linux. This patch ignores signal handlers that are
installed for signals that do not exist on the current platform.