Add support for skipping optional positions
This is useful in scenarios where you're able to infer a position's value from
elsewhere (e.g. a flag or environment variable), and want to leave the argument
for future positions.

This example covers most of the added functionality:

```ruby
class Opts < CLI::Kit::Opts
  def first
    # skip based on an option / flag
    position(skip: -> { skip? })
  end

  def second
    # skip based on the argument's value
    position(skip: ->(arg) { arg == 'skip' })
  end

  def skip?
    flag(long: '--skip')
  end
end
```

This also adds support for default values for optional positions, and unifies
some of the logic between options and positions.

We're also now able to put required arguments after optional ones, as it's
possible to skip over optional arguments and still satisfy required ones.