ErrorHandler: preserve exit status
This ensures it's possible to exit with a non-1 error status.

The error handler currently treats SystemExit a little too much
like "any other exception", meaning that calling `exit` with
a non-zero status causes the program to terminate with the
exception class printed to the terminal and a `1`.

This adjusts the exception handling to detect SystemExit a little
more cleanly, and pass through the specified exit status.
The special 30 exit status still keeps its old behaviour of
being translated to 1 before being termination.