While executing system commands poll IO pipes in a thread We are encountering an issue with Dev hanging while executing rails commands on apps using Spring. When executing the commands we open two pipes to pass as STDOUT and STDERR to the command process. We then poll in a loop until those streams are closed.
It appears that for some reason Spring does not properly close the STDERR stream, which causes Dev to hang waiting for the stream to be readable again. We are still investigating why this is happening in Spring.

However, Dev should be able to handle this situation more gracefully and not hang forever if a sub process is behaving badly. We can do a non-blocking check if the process has exited and break out of the loop if it is has.